Added WS routing for renegotiation answers
Switched to transmitting a stream since RTCPeerConnection changes the track ID when we add it, so we can't match them up on the other end. But stream IDs don't change strangely enough.

Moved the event handlers for RTCPeerConnection to just after creation and before we create the offer or add any tracks. This means we don't miss any ICE candidates now and if users are already streaming, we won't miss the track being added.